migrate gc tests to jvm
