+++ stone_multiplatform/build.gradle.kts
@@ -17,8 +17,10 @@ kotlin {
 
     linuxX64()
     mingwX64()
-    wasmJs()
-
+    wasmJs {
+        browser()
+        nodejs()
+    }
     sourceSets {
         commonMain.dependencies {
             api(project(":inject_multiplatform"))
