+++ weakref_multiplatform/src/wasmJsMain/kotlin/js/WeakRef.kt
@@ -1,5 +1,5 @@
 package js
 
-external class WeakRef<T : Any?>(target: T) {
+external class WeakRef<T : Any>(target: T) {
     fun deref(): T?
 }
